Move rollout reconstruction out of codex.rs
